Add training match rate logging for structure prediction
Compare generated validation samples against cached training samples to detect
memorization. Logs val/train_prim_match_rate metric showing fraction of generated
samples that match any training sample.

New config options in validation section:
- compute_train_match_rate: enable/disable feature (default: true)
- max_train_cache_size: max training samples to cache (default: 500)

Also fixes bfloat16 to numpy conversion error in virtual_coords_to_lattice_and_supercell.

+++ src/catgen/data/conversions.py
@@ -181,8 +181,9 @@ def virtual_coords_to_lattice_and_supercell(
     # Compute supercell matrix: supercell_matrix = supercell_vectors @ inv(prim_vectors)
     supercell_matrices = torch.zeros(B, 3, 3, device=device, dtype=dtype)
     for i in range(B):
-        prim_np = prim_virtual_coords[i].detach().cpu().numpy()
-        supercell_np = supercell_virtual_coords[i].detach().cpu().numpy()
+        # Convert to float32 for numpy compatibility (bfloat16 not supported)
+        prim_np = prim_virtual_coords[i].detach().cpu().float().numpy()
+        supercell_np = supercell_virtual_coords[i].detach().cpu().float().numpy()
         sm_np = compute_supercell_matrix_from_vectors(prim_np, supercell_np)
         supercell_matrices[i] = torch.from_numpy(sm_np).to(device=device, dtype=dtype)
